+++ packages/twenty-sdk/src/cli/commands/server.ts
@@ -1,49 +1,15 @@
 import { serverStart } from '@/cli/operations/server-start';
+import {
+  CONTAINER_NAME,
+  containerExists,
+  DEFAULT_PORT,
+  getContainerPort,
+  isContainerRunning,
+} from '@/cli/utilities/server/docker-container';
 import { checkServerHealth } from '@/cli/utilities/server/detect-local-server';
 import chalk from 'chalk';
 import type { Command } from 'commander';
 import { execSync, spawnSync } from 'node:child_process';
-
-const CONTAINER_NAME = 'twenty-app-dev';
-const DEFAULT_PORT = 2020;
-
-const isContainerRunning = (): boolean => {
-  try {
-    const result = execSync(
-      `docker inspect -f '{{.State.Running}}' ${CONTAINER_NAME}`,
-      { encoding: 'utf-8', stdio: ['pipe', 'pipe', 'ignore'] },
-    ).trim();
-
-    return result === 'true';
-  } catch {
-    return false;
-  }
-};
-
-const getContainerPort = (): number => {
-  try {
-    const result = execSync(
-      `docker inspect -f '{{(index (index .NetworkSettings.Ports "3000/tcp") 0).HostPort}}' ${CONTAINER_NAME}`,
-      { encoding: 'utf-8', stdio: ['pipe', 'pipe', 'ignore'] },
-    ).trim();
-
-    return parseInt(result, 10) || DEFAULT_PORT;
-  } catch {
-    return DEFAULT_PORT;
-  }
-};
-
-const containerExists = (): boolean => {
-  try {
-    execSync(`docker inspect ${CONTAINER_NAME}`, {
-      stdio: ['pipe', 'pipe', 'ignore'],
-    });
-
-    return true;
-  } catch {
-    return false;
-  }
-};
 
 const validatePort = (value: string): number => {
   const port = parseInt(value, 10);
