+++ packages/twenty-sdk/src/cli/operations/server-start.ts
@@ -1,63 +1,20 @@
+import { APP_ERROR_CODES, type CommandResult } from '@/cli/types';
 import { ConfigService } from '@/cli/utilities/config/config-service';
+import { runSafe } from '@/cli/utilities/run-safe';
+import {
+  checkDockerRunning,
+  CONTAINER_NAME,
+  containerExists,
+  DEFAULT_PORT,
+  getContainerPort,
+  IMAGE,
+  isContainerRunning,
+} from '@/cli/utilities/server/docker-container';
 import {
   checkServerHealth,
   detectLocalServer,
 } from '@/cli/utilities/server/detect-local-server';
-import { runSafe } from '@/cli/utilities/run-safe';
-import { APP_ERROR_CODES, type CommandResult } from '@/cli/types';
 import { execSync, spawnSync } from 'node:child_process';
-
-const CONTAINER_NAME = 'twenty-app-dev';
-const IMAGE = 'twentycrm/twenty-app-dev:latest';
-const DEFAULT_PORT = 2020;
-
-const isContainerRunning = (): boolean => {
-  try {
-    const result = execSync(
-      `docker inspect -f '{{.State.Running}}' ${CONTAINER_NAME}`,
-      { encoding: 'utf-8', stdio: ['pipe', 'pipe', 'ignore'] },
-    ).trim();
-
-    return result === 'true';
-  } catch {
-    return false;
-  }
-};
-
-const getContainerPort = (): number => {
-  try {
-    const result = execSync(
-      `docker inspect -f '{{(index (index .NetworkSettings.Ports "3000/tcp") 0).HostPort}}' ${CONTAINER_NAME}`,
-      { encoding: 'utf-8', stdio: ['pipe', 'pipe', 'ignore'] },
-    ).trim();
-
-    return parseInt(result, 10) || DEFAULT_PORT;
-  } catch {
-    return DEFAULT_PORT;
-  }
-};
-
-const containerExists = (): boolean => {
-  try {
-    execSync(`docker inspect ${CONTAINER_NAME}`, {
-      stdio: ['pipe', 'pipe', 'ignore'],
-    });
-
-    return true;
-  } catch {
-    return false;
-  }
-};
-
-const checkDockerRunning = (): boolean => {
-  try {
-    execSync('docker info', { stdio: 'ignore' });
-
-    return true;
-  } catch {
-    return false;
-  }
-};
 
 const HEALTH_POLL_INTERVAL_MS = 2000;
 const HEALTH_TIMEOUT_MS = 180 * 1000;
